back: refactored list and moderating controllers

+++ backend/controllers/list.controller.ts
@@ -1,4 +1,5 @@
 import { Request, Response } from 'express'
+import { decode } from 'jsonwebtoken'
 import { defaultError } from '../middleware/status-sender'
 import WaifusModel from '../models/list.model'
 
@@ -15,12 +16,27 @@ export const getWaifus = async (req, res) => {
 }
 
 export const removeWaifuById = async (req: Request, res: Response) => {
-  try {
-    await WaifusModel.findByIdAndRemove(req.body.id)
+  const { authorization } = req.headers
 
+  const token = decode(authorization.replace('Bearer ', '')) as Record<
+    string,
+    string
+  >
+
+  try {
+    const waifu = await WaifusModel.findById(req.body.id)
+    const waifuUserId = JSON.parse(JSON.stringify(waifu.user))
+    if (token.permission || waifuUserId === token._id) {
+      await WaifusModel.findByIdAndRemove(req.body.id)
+    } else {
+      return res.status(500).json({
+        text: 'You do not have permission',
+        color: 'warning'
+      })
+    }
     res.status(200).json({
       text: 'The post was successfully deleted',
-      color: 'color--success'
+      color: 'success'
     })
   } catch (e) {
     console.error(e)
